# [SPARK-31705][SQL] Push predicate through join by rewriting join condition to conjunctive normal form