@viirya 2020-05-20
I don't look at this PR yet, but yes, a general one should be better.